# Use hash-consing to optimise index cache