@dcoutts 2016-09-25
Interesting. The MUT time is up just a fraction (since we're doing more work to find the sharing) but the saving on GC time much larger.