@dcoutts 2016-09-25
Note this could be improved if the sharing were done incrementally as the cache is read in, rather than reading it all in and then sharing after. But that's tricky for the binary cache case I think since the binary decoder is not incremental.

@hvr could you add a TODO/NOTE pointing out that if/when we change the binary rep of the cache again, if we can make it decode incrementally (as we can with CBOR) then we can do the hash consing incrementally too, and reduce the peak mem use for loading the cache.